requirements.txt
flask==2.3.3
flask-cors==4.0.0
PyJWT==2.8.0
mysql-connector-python==8.1.0
python-dotenv==1.0.0
werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
redis==4.6.0